+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # [NEXT]
+- Render transform improvements.
+- Now the `SimpleDirectionAnimation` do flip component that use it as necessary.
+- Adds param `useCompFlip` in `playOnce` (default `false`). If `true` the animation is flipped equal current state of component.
 - Adds param `backgroundColor` in `BonfireWidget`.
+- create `BarLifeComponent`.
+- Adds `UseBarLife` mixin.
+- method `drawDefaultLifeBar` now is deprecated. Pls use `UseBarLife` mixin.
 
 # [2.10.10]
 - Update Flame to `1.4.0`.